feat: Implement shortcut management system with Zustand
- Add `useShortcutStore` for managing keyboard shortcuts and extensions.
- Create types for `ShortcutAction`, `ShortcutExtension`, and `ShortcutStore`.
- Implement `FormatShortcutExtension`, `MarkShortcutExtension`, and `MarkdownShortcutExtension` for formatting actions.
- Introduce slash command functionality with `OpenSlashCommandShortcutExtension` and `SlashCommandShortcutExtension`.
- Develop utility functions for handling slash command menu operations.
- Create Zustand store for managing slash menu state and actions.
- Refactor existing code to integrate new shortcut management system.
