+++ src/core/shortcut/index.ts
@@ -1,6 +1,5 @@
-export { useShortcutStore } from "./store/shortcutStore.ts";
+export { useShortcutStore } from "./store/useShortcutStore.ts";
 export { default as useRegisterShortcuts } from "./hooks/useRegisterShortcuts.ts";
-export type {
-  ShortcutExtension as ShortcutPlugin,
-  ShortcutAction,
-} from "./store/shortcutStore.ts";
+export { type default as ShortcutAction } from "./types/ShortcutAction.ts";
+export { type  default as ShortcutExtension } from "./types/ShortcutExtension.ts";
+export { type default as ShortcutStore } from "./types/ShortcutStore.ts";
