+++ src/features/editor/plugins/format/MarkShortcutExtension.ts
@@ -1,6 +1,7 @@
-import { type ShortcutExtension } from "@/core/shortcut/store/shortcutStore";
+import type { ShortcutExtension } from "@/core/shortcut";
+import type { Editor } from "slate";
 
-const MarkShortcutExtension: ShortcutExtension = {
+const MarkShortcutExtension: ShortcutExtension<Editor> = {
   name: "mark",
   priority: 10,
   actions: {
