+++ src/features/editor/plugins/slash-menu/handleOpenSlashMenu.ts
